+++ include/SQLiteCpp/Database.h
@@ -120,7 +120,7 @@ public:
              const int          aBusyTimeoutMs  = 0,
              const std::string& aVfs            = "");
 
-#if __cplusplus >= 201103L
+#if __cplusplus >= 201103L || (defined(_MSC_VER) && _MSC_VER >= 1600)
     /**
      * @brief Move an SQLite database connection.
      *