+++ tests/Database_test.cpp
@@ -29,14 +29,16 @@ void assertion_failed(const char* apFile, const long apLine, const char* apFunc,
 }
 #endif
 
-TEST(SQLiteCpp, version) {
+TEST(SQLiteCpp, version)
+{
     EXPECT_STREQ(SQLITE_VERSION,        SQLite::VERSION);
     EXPECT_EQ   (SQLITE_VERSION_NUMBER, SQLite::VERSION_NUMBER);
     EXPECT_STREQ(SQLITE_VERSION,        SQLite::getLibVersion());
     EXPECT_EQ   (SQLITE_VERSION_NUMBER, SQLite::getLibVersionNumber());
 }
 
-TEST(Database, ctorExecCreateDropExist) {
+TEST(Database, ctorExecCreateDropExist)
+{
     remove("test.db3");
     {
         // Try to open a non-existing database
@@ -70,7 +72,8 @@ SQLite::Database DatabaseBuilder(const char* apName)
     return SQLite::Database(apName, SQLite::OPEN_READWRITE | SQLite::OPEN_CREATE);
 }
 
-TEST(Database, moveConstructor) {
+TEST(Database, moveConstructor)
+{
     remove("test.db3");
     {
         // Create a new database, using the move constructor
@@ -87,7 +90,8 @@ TEST(Database, moveConstructor) {
 
 #endif 
 
-TEST(Database, createCloseReopen) {
+TEST(Database, createCloseReopen)
+{
     remove("test.db3");
     {
         // Try to open the non-existing database
@@ -107,7 +111,8 @@ TEST(Database, createCloseReopen) {
     remove("test.db3");
 }
 
-TEST(Database, inMemory) {
+TEST(Database, inMemory)
+{
     {
         // Create a new database
         SQLite::Database db(":memory:", SQLite::OPEN_READWRITE);
@@ -126,7 +131,8 @@ TEST(Database, inMemory) {
 }
 
 #if SQLITE_VERSION_NUMBER >= 3007015 // SQLite v3.7.15 is first version with PRAGMA busy_timeout
-TEST(Database, busyTimeout) {
+TEST(Database, busyTimeout)
+{
     {
         // Create a new database with default timeout of 0ms
         SQLite::Database db(":memory:");
@@ -163,7 +169,8 @@ TEST(Database, busyTimeout) {
 }
 #endif // SQLITE_VERSION_NUMBER >= 3007015
 
-TEST(Database, exec) {
+TEST(Database, exec)
+{
     // Create a new database
     SQLite::Database db(":memory:", SQLite::OPEN_READWRITE);
 
@@ -224,7 +231,8 @@ TEST(Database, exec) {
 #endif
 }
 
-TEST(Database, execAndGet) {
+TEST(Database, execAndGet)
+{
     // Create a new database
     SQLite::Database db(":memory:", SQLite::OPEN_READWRITE);
 
@@ -242,7 +250,8 @@ TEST(Database, execAndGet) {
     EXPECT_EQ(3,            db.execAndGet("SELECT weight FROM test WHERE value=\"first\"").getInt());
 }
 
-TEST(Database, execException) {
+TEST(Database, execException)
+{
     // Create a new database
     SQLite::Database db(":memory:", SQLite::OPEN_READWRITE);
     EXPECT_EQ(SQLite::OK, db.getErrorCode());
@@ -284,7 +293,8 @@ TEST(Database, execException) {
 // TODO: test Database::loadExtension()
 
 #ifdef SQLITE_HAS_CODEC
-TEST(Database, encryptAndDecrypt) {
+TEST(Database, encryptAndDecrypt)
+{
     remove("test.db3");
     {
         // Try to open the non-existing database
@@ -328,7 +338,8 @@ TEST(Database, encryptAndDecrypt) {
     remove("test.db3");
 }
 #else // SQLITE_HAS_CODEC
-TEST(Database, encryptAndDecrypt) {
+TEST(Database, encryptAndDecrypt)
+{
     remove("test.db3");
     {
         // Try to open the non-existing database