+++ CHANGELOG.md
@@ -130,6 +130,7 @@ Version 2.3.0 - March 3 2019
 - Add implicit cast operator to char and short types #179 #180
 
 Version ?
+- Update SQLite3 from 3.27.2 to 3.28.0 (2019-04-16)
 - #191 CMake Warning line 299
 - #190 Implement move constructors
 - #192 Add wrapper for bind parameter count